+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.0.1] - unreleased
+
+### Fixed
+- Pre-filter CMS content before it is passed to Summernote editor
+
+### Added
+- `HtmlTagRemover` class
+
 ## [4.0.0] - 2024-03-12
 
 ### Changed